Byt färg på text.
